Module tetka_games::ataxx

Structs

BitBoard

ColorIter

An iterator over the variants of Color

ColoredPieceIter

An iterator over the variants of ColoredPiece

FileIter

An iterator over the variants of File

Move

Move represents an Ataxx move which can be played on the Board.

PieceIter

An iterator over the variants of Piece

Position

Position represents the snapshot of an Ataxx Board, the state of the an ataxx game at a single point in time. It also provides all of the methods necessary to manipulate such a snapshot.

RankIter

An iterator over the variants of Rank

SquareIter

An iterator over the variants of Square

Enums

Color

Color represents all the possible colors that an ataxx piece can have, specifically, Black and White.

ColoredPiece

Piece represents all the possible ataxx pieces.

File

File represents a file on the Ataxx Board. Each vertical column of Squares on an Ataxx Board is known as a File. There are 7 of them in total.

MoveParseError

Piece

Piece represents the types of pieces in ataxx, namely Piece and Block.

PositionParseError

PositionParseErr represents an error encountered while parsing the given FEN position field into a valid Position.

Rank

Rank represents a rank on the Ataxx Board. Each horizontal row of Squares on an Ataxx Board is known as a Rank. There are 7 of them in total.

Square

Square represents all the squares present on an Ataxx Board. The index of each Square is equal to rank-index * 8 + file-index.