Enum tetka_games::ataxx::Square

#[repr(u8)]
pub enum Square {

    A1 = 0,
    B1 = 1,
    C1 = 2,
    D1 = 3,
    E1 = 4,
    F1 = 5,
    G1 = 6,
    A2 = 7,
    B2 = 8,
    C2 = 9,
    D2 = 10,
    E2 = 11,
    F2 = 12,
    G2 = 13,
    A3 = 14,
    B3 = 15,
    C3 = 16,
    D3 = 17,
    E3 = 18,
    F3 = 19,
    G3 = 20,
    A4 = 21,
    B4 = 22,
    C4 = 23,
    D4 = 24,
    E4 = 25,
    F4 = 26,
    G4 = 27,
    A5 = 28,
    B5 = 29,
    C5 = 30,
    D5 = 31,
    E5 = 32,
    F5 = 33,
    G5 = 34,
    A6 = 35,
    B6 = 36,
    C6 = 37,
    D6 = 38,
    E6 = 39,
    F6 = 40,
    G6 = 41,
    A7 = 42,
    B7 = 43,
    C7 = 44,
    D7 = 45,
    E7 = 46,
    F7 = 47,
    G7 = 48,

}

Square represents all the squares present on an Ataxx Board. The index of each Square is equal to rank-index * 8 + file-index.

Variants

A1 = 0

B1 = 1

C1 = 2

D1 = 3

E1 = 4

F1 = 5

G1 = 6

A2 = 7

B2 = 8

C2 = 9

D2 = 10

E2 = 11

F2 = 12

G2 = 13

A3 = 14

B3 = 15

C3 = 16

D3 = 17

E3 = 18

F3 = 19

G3 = 20

A4 = 21

B4 = 22

C4 = 23

D4 = 24

E4 = 25

F4 = 26

G4 = 27

A5 = 28

B5 = 29

C5 = 30

D5 = 31

E5 = 32

F5 = 33

G5 = 34

A6 = 35

B6 = 36

C6 = 37

D6 = 38

E6 = 39

F6 = 40

G6 = 41

A7 = 42

B7 = 43

C7 = 44

D7 = 45

E7 = 46

F7 = 47

G7 = 48