Struct tetka_games::ataxx::BitBoard

pub struct BitBoard(pub u64);

Tuple Fields
0: u64
Implementations

impl BitBoard

pub fn singles(bb: BitBoard) -> BitBoard

pub const fn single(square: Square) -> BitBoard

single returns the targets of a singular Move from the given Square.

pub const fn double(square: Square) -> BitBoard

double returns the targets of a jump Move from the given Square.