Module tetka_games::interface

Structs

Hash

Hash represents the semi-unique checksum of a Position used to efficiently check for Position equality. Some properties that a Hash should possess include determinism, uniform distribution, avalanche effect, and collision resistance.

Enums

PiecePlacementParseError

PositionParseErr represents an error encountered while parsing the given FEN position field into a valid Position.

TypeParseError

Traits

BitBoardType

BitBoardType is a generalized interface implemented by BitBoards of arbitrary size. This allows programs to handle BitBoards of any size with generic functions using this common interface.

ColoredPieceType

The ColoredPiece trait should be implemented by the piece representation (with color) for a game.

MoveStore

MoveStore is a trait implemented by types which are able to store moves inside themselves and are thus usable in move-generation methods in Position like generate_moves_into<T>.

MoveType

The Move trait should be implemented the move representation of a game.

PositionType

Position is a generalized interface for board representations of a wide range of games. It can be used to create game-agnostic software. Tetka provides some of the popular board representations out of the box, but custom ones can also be implemented by the library user.

RepresentableType

RepresentableType is a basic trait which is implemented by enums with both a binary and string representation and backed by an integer.

SquareType

SquareType is the trait implemented by the type for Squares in a game’s board representation. It provides a generic interface along with useful predefined methods that a Square type would require.

Type Aliases

MoveList

MoveList is a basic implementation of MoveStore that is used to allow users to utilize move-generation methods without having to implement a MoveStore by themselves. It also has utility methods other than the MoveStore trait.