Trait tetka_games::interface::BitBoardType

pub trait BitBoardType: Sized + Copy + Eq + Into<Self::Base> + From<Self::Square> + Not<Output = Self> + Sub<usize, Output = Self> + Shr<usize, Output = Self> + Shl<usize, Output = Self> + BitOr<Self, Output = Self> + BitAnd<Self, Output = Self> + BitXor<Self, Output = Self> + Iterator<Item = Self::Square>
where
    Self::Base: PrimInt,
    Self::Square: SquareType,
{
    type Base;
    type Square;

    const EMPTY: Self;
    const UNIVERSE: Self;
    const FIRST_FILE: Self;
    const FIRST_RANK: Self;

    // Provided methods
    fn new() -> Self { ... }
    fn is_disjoint(self, other: Self) -> bool { ... }
    fn is_subset(self, other: Self) -> bool { ... }
    fn is_superset(self, other: Self) -> bool { ... }
    fn is_empty(self) -> bool { ... }
    fn len(self) -> usize { ... }
    fn contains(self, square: Self::Square) -> bool { ... }
    fn north(self) -> Self { ... }
    fn south(self) -> Self { ... }
    fn east(self) -> Self { ... }
    fn west(self) -> Self { ... }
    fn insert(&mut self, square: Self::Square) { ... }
    fn remove(&mut self, square: Self::Square) { ... }
    fn pop_lsb(&mut self) -> Option<Self::Square> { ... }
    fn pop_msb(&mut self) -> Option<Self::Square> { ... }
    fn clear(&mut self) { ... }
    fn lsb(self) -> Option<Self::Square> { ... }
    fn msb(self) -> Option<Self::Square> { ... }
    fn retain<F: FnMut(Self::Square) -> bool>(&mut self, f: F) { ... }
    fn file(file: <Self::Square as SquareType>::File) -> Self { ... }
    fn rank(rank: <Self::Square as SquareType>::Rank) -> Self { ... }

}

BitBoardType is a generalized interface implemented by BitBoards of arbitrary size. This allows programs to handle BitBoards of any size with generic functions using this common interface.

Required Associated Types

type Base

type Square

Required Associated Constants

const EMPTY: Self

EMPTY is an empty Self containing no Squares.

const UNIVERSE: Self

UNIVERSE is a filled Self containing all Squares.

const FIRST_FILE: Self

const FIRST_RANK: Self

Provided Methods

fn new() -> Self

Makes a new, empty BitBoard.

fn is_disjoint(self, other: Self) -> bool

Returns true if self has no elements in common with other. This is equivalent to checking for an empty intersection.

fn is_subset(self, other: Self) -> bool

Returns true if the BitBoard is a subset of another, i.e., other contains at least all the values in self.

fn is_superset(self, other: Self) -> bool

Returns true if the BitBoard is a superset of another, i.e., self contains at least all the values in other.

fn is_empty(self) -> bool

Returns true if the BitBoard contains no elements.

fn len(self) -> usize

Returns the number of elements in the BitBoard.

fn contains(self, square: Self::Square) -> bool

Returns true if the BitBoard contains a value.

fn north(self) -> Self

north returns a new Self with all the squares shifted to the north.

fn south(self) -> Self

south returns a new Self with all the squares shifted to the south.

fn east(self) -> Self

east returns a new Self with all the squares shifted to the east.

fn west(self) -> Self

west returns a new Self with all the squares shifted to the west.

fn insert(&mut self, square: Self::Square)

Adds square to the BitBoard.

fn remove(&mut self, square: Self::Square)

Removes square from the BitBoard.

fn pop_lsb(&mut self) -> Option<Self::Square>

pop_lsb pops the least significant Self::Square from the Self, i.e. it removes the lsb from the Self and returns its value.

fn pop_msb(&mut self) -> Option<Self::Square>

pop_msb pops the most significant Self::Square from the Self i.e. it removes the msb from the Self and returns its value.

fn clear(&mut self)

Clears the BitBoard, removing all elements.

fn lsb(self) -> Option<Self::Square>

get_lsb returns the least significant Self::Square from the Self.

fn msb(self) -> Option<Self::Square>

get_msb returns the most significant Self::Square from the Self.

fn retain<F: FnMut(Self::Square) -> bool>(&mut self, f: F)

Retains only the elements specified by the predicate.

In other words, remove all elements s for which f(s) returns false. The elements are visited in ascending order.

fn file(file: <Self::Square as SquareType>::File) -> Self

Returns a BitBoard containing all the squares from the given File.

fn rank(rank: <Self::Square as SquareType>::Rank) -> Self

Returns a BitBoard containing all the squares from the given Rank.

Object Safety

This trait is not object safe.