Trait tetka_games::interface::MoveType

pub trait MoveType: FromStr + Display + From<u16> + Into<u16> + Copy {
    const NULL: Self;
    const MAX_IN_GAME: usize;
    const MAX_IN_POSITION: usize;
}

The Move trait should be implemented the move representation of a game.

Required Associated Constants

const NULL: Self

NULL represents the null or the ‘do nothing’ move.

const MAX_IN_GAME: usize

MAX_IN_GAME is a suitably high maximum for the number of move in a game.

const MAX_IN_POSITION: usize

MAX_IN_POSITION is a suitably high maximum for the number of move in a single, possibly unreachable position.

Object Safety

This trait is not object safe.