Trait tetka_games::interface::PositionType

pub trait PositionType: FromStr + Display
where
    Self::BitBoard: BitBoardType,
    Self::ColoredPiece: ColoredPieceType,
    Self::Move: MoveType,
{
    type BitBoard;
    type ColoredPiece;
    type Move;

    // Required methods
    fn insert(
        &mut self,
        sq: <Self::BitBoard as BitBoardType>::Square,
        piece: Self::ColoredPiece,
    );
    fn remove(
        &mut self,
        sq: <Self::BitBoard as BitBoardType>::Square,
    ) -> Option<Self::ColoredPiece>;
    fn at(
        &self,
        sq: <Self::BitBoard as BitBoardType>::Square,
    ) -> Option<Self::ColoredPiece>;
    fn piece_bb(
        &self,
        piece: <Self::ColoredPiece as ColoredPieceType>::Piece,
    ) -> Self::BitBoard;
    fn color_bb(
        &self,
        color: <Self::ColoredPiece as ColoredPieceType>::Color,
    ) -> Self::BitBoard;
    fn colored_piece_bb(&self, piece: Self::ColoredPiece) -> Self::BitBoard;
    fn hash(&self) -> Hash;
    fn winner(&self) -> Option<<Self::ColoredPiece as ColoredPieceType>::Color>;
    fn after_move<const UPDATE_PERIPHERALS: bool>(
        &self,
        mov: Self::Move,
    ) -> Self;
    fn generate_moves_into<const ALLOW_ILLEGAL: bool, const QUIET: bool, const NOISY: bool, T: MoveStore<Self::Move>>(
        &self,
        movelist: &mut T,
    );

    // Provided methods
    fn is_game_over(&self) -> bool { ... }
    fn generate_moves<const ALLOW_ILLEGAL: bool, const QUIET: bool, const NOISY: bool>(
        &self,
    ) -> MoveList<Self::Move> { ... }
    fn count_moves<const QUIET: bool, const NOISY: bool>(&self) -> usize { ... }

}

Position is a generalized interface for board representations of a wide range of games. It can be used to create game-agnostic software. Tetka provides some of the popular board representations out of the box, but custom ones can also be implemented by the library user.

Required Associated Types

type BitBoard

The type for the bitboards used by this board representation.

type ColoredPiece

The type for the pieces (with color) used by this board representation.

type Move

The type for one move in this board representation.

Required Methods

fn insert( &mut self, sq: <Self::BitBoard as BitBoardType>::Square, piece: Self::ColoredPiece, )

insert puts the given piece at the given square. The implementation is free to assume that the square is currently empty.

fn remove( &mut self, sq: <Self::BitBoard as BitBoardType>::Square, ) -> Option<Self::ColoredPiece>

remove clears the given square, and returns the piece that was there.

fn at( &self, sq: <Self::BitBoard as BitBoardType>::Square, ) -> Option<Self::ColoredPiece>

at returns the piece that is in the given square.

fn piece_bb( &self, piece: <Self::ColoredPiece as ColoredPieceType>::Piece, ) -> Self::BitBoard

fn color_bb( &self, color: <Self::ColoredPiece as ColoredPieceType>::Color, ) -> Self::BitBoard

fn colored_piece_bb(&self, piece: Self::ColoredPiece) -> Self::BitBoard

fn hash(&self) -> Hash

fn winner(&self) -> Option<<Self::ColoredPiece as ColoredPieceType>::Color>

winner returns the winning side in the current position.

fn after_move<const UPDATE_PERIPHERALS: bool>(&self, mov: Self::Move) -> Self

after_move returns the position after playing the given move on the current position. The UPDATE_PERIPHERALS flag can be interpreted as toggling the non-essential updated which are done by this function, like the hash function for the position.

fn generate_moves_into<const ALLOW_ILLEGAL: bool, const QUIET: bool, const NOISY: bool, T: MoveStore<Self::Move>>( &self, movelist: &mut T, )

generate_moves_into generates all the moves in the current position into the given move storage. The QUIET and NOISY flags toggles the generation of reversible and irreversible moves respectively.

Provided Methods

fn is_game_over(&self) -> bool

is_game_over returns a boolean representing if the game is over.

fn generate_moves<const ALLOW_ILLEGAL: bool, const QUIET: bool, const NOISY: bool>( &self, ) -> MoveList<Self::Move>

generate_moves is similar to generate_moves_into, except that instead of taking some storage as input it stores into a custom stack-based type.

fn count_moves<const QUIET: bool, const NOISY: bool>(&self) -> usize

count_moves is similar to generate_moves, except instead of returning a list of the available moves, it returns the number of available moves. By default this is simply generate_moves().len(), but implementations may take advantage of various optimizations counting as opposed to storing the moves allows to provide a more efficient version.

Object Safety

This trait is not object safe.