Trait tetka_games::interface::RepresentableType

pub trait RepresentableType<B: Into<usize>>: Copy + Eq + FromStr + Display + Into<B> + TryFrom<B, Error: Debug> + IntoEnumIterator {
    const N: usize;

    // Provided method
    unsafe fn unsafe_from<T: Copy + Into<usize>>(number: T) -> Self { ... }
}

RepresentableType is a basic trait which is implemented by enums with both a binary and string representation and backed by an integer.

Required Associated Constants

const N: usize

N is the number of specializations of the enum.

Provided Methods

unsafe fn unsafe_from<T: Copy + Into<usize>>(number: T) -> Self

unsafe_from unsafely converts the given number into Self.

Safety

unsafe_from assumes that the target type can represent the provided number, i.e. the number has a valid representation in the target type. The function comes with a debug check for the same, and failure to uphold this invariant will result in undefined behavior.

Object Safety

This trait is not object safe.