Trait tetka_games::interface::SquareType

pub trait SquareType: RepresentableType<u8>
where
    Self::File: RepresentableType<u8>,
    Self::Rank: RepresentableType<u8>,
{
    type File;
    type Rank;

    // Provided methods
    fn new(file: Self::File, rank: Self::Rank) -> Self { ... }
    fn file(self) -> Self::File { ... }
    fn rank(self) -> Self::Rank { ... }
    fn north(self) -> Option<Self> { ... }
    fn south(self) -> Option<Self> { ... }
    fn east(self) -> Option<Self> { ... }
    fn west(self) -> Option<Self> { ... }
}

SquareType is the trait implemented by the type for Squares in a game’s board representation. It provides a generic interface along with useful predefined methods that a Square type would require.

Required Associated Types

type File

type Rank

Provided Methods

fn new(file: Self::File, rank: Self::Rank) -> Self

new creates a new Square from the given File and Rank.

fn file(self) -> Self::File

Returns the File of self.

fn rank(self) -> Self::Rank

Returns the Rank of self.

fn north(self) -> Option<Self>

Returns the square to the north of self. If there is no Square to the north of self, it returns None.

fn south(self) -> Option<Self>

Returns the square to the south of self. If there is no Square to the south of self, it returns None.

fn east(self) -> Option<Self>

Returns the square to the east of self. If there is no Square to the east of self, it returns None.

fn west(self) -> Option<Self>

Returns the square to the west of self. If there is no Square to the west of self, it returns None.

Object Safety

This trait is not object safe.